Restructure base into module
