+++ lib/base/index.js
@@ -1,8 +1,8 @@
 // @flow
-import {keybaseExec} from './utils'
-import {withStore, actionInit} from './store'
-import type {Store} from './store/types'
-import type {InitOptions, DeviceUsernamePair, UsernameAndDevice} from './types.js'
+import {keybaseExec} from '../utils'
+import {withStore, actionInit} from '../store'
+import type {Store} from '../store/types'
+import type {InitOptions, DeviceUsernamePair, UsernameAndDevice} from '../types.js'
 
 /**
  * @ignore
