+++ lib/bot.js
@@ -1,5 +1,5 @@
 // @flow
-import Base from './base-bot'
+import Base from './base'
 import Chat from './chat'
 import {createStore, reducer, initialState} from './store'
 import type {Store} from './store/types'
