+++ lib/chat/index.js
@@ -1,5 +1,5 @@
 // @flow
-import {guardInitialized} from '../base-bot'
+import {guardInitialized} from '../base'
 import {formatOptions, keybaseExec} from '../utils'
 import {CHAT_API_VERSION} from '../constants'
 import {withStore, withGuards} from '../store'
